requirements.txt
pillow
customtkinter
google-generativeai